Fix TArray<UObject> parsing error
* Move parsing logic to ModStatic
* Omit LogLevel::Verbose in LogOutput
