+++ src/statics.h
@@ -2,17 +2,29 @@
 
 #include <string>
 #include <DynamicOutput/Output.hpp>
+#include <Unreal/UnrealCoreStructs.hpp>
+#include <LuaMadeSimple/LuaMadeSimple.hpp>
 
-constexpr std::wstring logLevelToString(RC::LogLevel::LogLevel level) {
+using namespace RC;
+using namespace RC::Unreal;
+using namespace RC::LuaMadeSimple;
+
+constexpr std::wstring logLevelToString(LogLevel::LogLevel level) {
 	switch (level) {
-	case RC::LogLevel::Default: return L"DEFAULT";
-	case RC::LogLevel::Normal: return L"NORMAL";
-	case RC::LogLevel::Verbose: return L"VERBOSE";
-	case RC::LogLevel::Warning: return L"WARNING";
-	case RC::LogLevel::Error: return L"ERROR";
+	case LogLevel::Default: return L"DEFAULT";
+	case LogLevel::Normal: return L"NORMAL";
+	case LogLevel::Verbose: return L"VERBOSE";
+	case LogLevel::Warning: return L"WARNING";
+	case LogLevel::Error: return L"ERROR";
 	default: return L"UNKNOWN";
 	}
 }
+
+enum PropertyType {
+	None = 0,
+	Array,
+	Map,
+};
 
 class ModStatics
 {
@@ -30,7 +42,7 @@ public:
 	static const std::string GetWebhookUrl();
 
 	// Primary template for the wrapper function
-	template<RC::LogLevel::LogLevel Level, typename ...Args>
+	template<LogLevel::LogLevel Level = LogLevel::Verbose, typename ...Args>
 	inline static auto LogOutput(std::wstring format, Args ...args) -> void
 	{
 		// Create the formatted message using ostringstream to avoid std::format issues
@@ -38,7 +50,9 @@ public:
 		if constexpr (sizeof...(args) > 0) {
 			try {
 				// Use vformat with make_wformat_args for runtime formatting
-				formatted_message = fmt::vformat(fmt::detail::to_string_view(format), fmt::make_format_args<fmt::buffer_context<wchar_t>>(args...));
+				formatted_message = fmt::vformat(
+					fmt::detail::to_string_view(format), 
+					fmt::make_format_args<fmt::buffer_context<wchar_t>>(args...));
 			}
 			catch (...) {
 				// Fallback if formatting fails
@@ -59,6 +73,19 @@ public:
 		// Create the prefixed format string
 		std::wstring prefixed_format = L"[" + GetModName() + L"] " + logLevelToString(Level) + L": " + formatted_message + L"\n";
 
-		RC::Output::send<Level>(prefixed_format);
+		Output::send<Level>(prefixed_format);
 	}
+
+	/**
+	* Convert a property to a Lua table
+	* @param property Reference to the pointer to export
+	* @param data Outer property object, usually the container of the property
+	* @param table Lua table to place the property into
+	*/
+	static void ExportPropertyAsTable(
+		FProperty* property,
+		void* data,
+		Lua::Table& table,
+		const PropertyType propertyType = PropertyType::None,
+		const bool convertObject = false);
 };
